backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
geoalchemy2>=0.14.3
alembic>=1.13.1
psycopg2-binary>=2.9.9
greenlet>=3.0.3
pgvector>=0.2.4

# Redis and Celery
redis>=5.0.1
celery>=5.3.6

# HTTP clients
httpx>=0.26.0
aiohttp>=3.9.0

# AI
anthropic>=0.18.0
openai>=1.12.0

# Geospatial
shapely>=2.0.2
pyproj>=3.6.1
geopy>=2.4.1
polyline>=2.0.0
gpxpy>=1.6.2

# Data processing
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
numpy>=1.26.0
orjson>=3.9.0

# Web scraping (for research service)
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utils
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
